+++ old.ts
@@ -288,57 +288,7 @@ function addEventListeners() {
 
   });
 
-  /**
-   * Event listeners for slingshot mode (desktop and mobile)
-   * Holding mouse down starts creating slingshot line
-   * Releasing mouse
-   */
-  canvas.addEventListener("mousedown", (e) => {
-    if (!slingshotInputElement.checked) {
-      return
-    }
-    this.xSlingshotStart = e.clientX;
-    this.ySlingshotStart = e.clientY;
-    xLineStart = e.clientX;
-    yLineStart = e.clientY;
-    startLine = true;
-  });
 
-  canvas.addEventListener("mousemove", (e) => {
-    if (!slingshotInputElement.checked) {
-      return
-    }
-    if (startLine === true) {
-      line = new Line(xLineStart, yLineStart, e.clientX, e.clientY);
-    }
-  });
-
-  canvas.addEventListener("mouseup", (e) => {
-    if (!slingshotInputElement.checked) {
-      return
-    }
-
-    let xPos = e.clientX;
-    let yPos = e.clientY;
-
-    let xPosEnd = e.clientX;
-    let yPosEnd = e.clientY;
-
-    startLine = false;
-    line = null;
-
-    /**
-     * Velocities increase the more you drag the slingshot
-     * Slingshot start positions are set on mouse down and used here
-     * @type {number}
-     * @type {number}
-     */
-    let xVelocity = (this.xSlingshotStart - xPosEnd) / 5;
-    let yVelocity = (this.ySlingshotStart - yPosEnd) / 5;
-
-    addNewBall(xPos, yPos, xVelocity, yVelocity);
-
-  });
 
   /**
    * Add click listener for clear
