--- src/features/slingshotMode.ts
@@ -1,13 +1,0 @@
-export function addMouseDragListener() {
-
-}
-
-export default {
-  isOn() {
-    const slingshotModeEnabled = ""
-    return slingshotModeEnabled
-  },
-
-  drawSlingshotLineIfActive() {
-  }
-}
