+++ src/lib/settings.ts
@@ -5,3 +5,10 @@ export function adjustForRetinaDisplay(value: number) {
   const RETINA_DISPLAY_MULTIPLIER = 2;
   return value * RETINA_DISPLAY_MULTIPLIER
 }
+
+const slingshotCheckbox = <HTMLInputElement>document.getElementById("slingshot");
+export const slingshotMode = {
+  isOn() {
+    return slingshotCheckbox.checked;
+  }
+}
